+++ NEWS.md
@@ -1,3 +1,5 @@
+# recipes (development version)
+
 # recipes 1.3.1
 
 * Fixed bug where `tune_args()` would error if argument to step had a parsnip object with tuned arguments. (#1506)