+++ NEWS.md
@@ -1,4 +1,4 @@
-# recipes (development version)
+# recipes 1.3.2
 
 * `step_nnmf()` is deprecated. Use `step_nnmf_sparse()` instead. (#1497)